+++ source/Messaging.Infrastructure/Configuration/MessageBus/IServiceBusClientConfiguration.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Messaging.Infrastructure.Transactions.MoveIn;
+namespace Messaging.Infrastructure.Configuration.MessageBus;
 
 /// <summary>
 /// Interface for configurations
